best lap time completed

+++ Mobile Car Racing Game/Assets/Scripts/lapCompletion.cs
@@ -65,7 +65,7 @@ public class lapCompletion : MonoBehaviour
             }
         //}
 
-        millSecondDisplay.GetComponent<TextMeshProUGUI>().text = "" + lapTimeManager.milliSecondCount;
+        millSecondDisplay.GetComponent<TextMeshProUGUI>().text = "" + (int)lapTimeManager.milliSecondCount;
 
         //PlayerPrefs.SetInt("MinuteSave", lapTimeManager.minuteCount);
         //PlayerPrefs.SetInt("SecondSave", lapTimeManager.secondCount);
